+++ InTouch/app/src/main/java/com/n8/intouch/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.n8.intouch
+package com.n8.intouch.main
 
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
@@ -8,6 +8,8 @@ import com.firebase.client.DataSnapshot
 import com.firebase.client.Firebase
 import com.firebase.client.FirebaseError
 import com.firebase.client.ValueEventListener
+import com.n8.intouch.InTouchApplication
+import com.n8.intouch.R
 import javax.inject.Inject
 import javax.inject.Named
